Scraping/requirements.txt
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
slack_sdk==3.39.0
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
